requirements.txt
pip~=23.1.2
setuptools~=65.5.0
Jinja2~=3.1.2
MarkupSafe~=2.1.3
packaging~=23.1
pytest~=7.3.2
pluggy~=1.0.0
iniconfig~=2.0.0
PyYAML~=6.0